+++ src/hcex/hcex_prepare_halo_render.cpp
@@ -17,13 +17,13 @@ extern "C" rasterizer_window_begin_parameters global_window_parameters;
 extern "C" int16_t global_current_render_target;
 extern "C" render_window hcex_window[];
 
-extern "C" int16_t hcex_render_player_index(int cam_idx);
+extern "C" int hcex_render_player_index(int cam_idx);
 extern "C" void render_camera_build_frustum_bounds(const render_camera *camera, real_rectangle2d *frustum_bounds);
 extern "C" void render_camera_build_frustum(const render_camera *camera, const real_rectangle2d *frustum_bounds, render_frustum *frustum, uint8_t build_projection);
 
 extern "C" void hcex_prepare_halo_render(int cam_idx)
 {
-    int16_t local_player_index = hcex_render_player_index(cam_idx);
+    int local_player_index = hcex_render_player_index(cam_idx); /* DEVIATION: was int16_t; no extsh after the call (mr r28,r3 @0x8369E6B8), cmpwi r28,-1 @0x8369E724 tests the full word; only the sth @0x8369E718 narrows */
     memset(&global_window_parameters, 0, sizeof(global_window_parameters));
 
     render_camera *rasterizer_camera = &hcex_window[cam_idx + 1].rasterizer_camera;
